Unify symbol-to-name logic
